Add Client::System::Input::InputData
